move to headers, start doing main script and whatnot
