+++ src/motors.cpp
@@ -1,11 +1,11 @@
-#include <motors.h>
+#include "headers/motors.h"
 
 /// @brief Moves an individual motor.
 /// @param pwm PWM servo driver instance.
 /// @param motor_no Motor number.
 /// @param degrees Degrees to offset.
 void moveMotor(Adafruit_PWMServoDriver &pwm, uint8_t motor_no, float degrees) {
-    // TODO: move one individual motor by `degrees`
+    // TODO: move one individual motor by `degrees`. Waiting on encoder support and interfacing.
 }
 
 /// @brief Moves all three motors by angles `delta_angles`.
@@ -24,4 +24,25 @@ BLA::Matrix<3> getMotorAngles() {
     // Waiting on encoder support and interfacing.
 }
 
-// TODO: Implement moveToPosition() in the similar way that we would call move_to_position() for the LEGO motors.
+/// @brief Homes a singular motor.
+/// @param pwm PWM controller instance.
+/// @param encoder Corresponding encoder instance to the motor.
+/// @param indexPin Index pin of the encoder.
+/// @param pwmChannel Motor channel number. Identifies the motor number.
+void homeMotor(Adafruit_PWMServoDriver &pwm, Encoder& encoder, int indexPin, int pwmChannel) {
+    pinMode(indexPin, INPUT);
+
+    // Rotate motor slowly in one direction
+    pwm.setPWM(pwmChannel, 0, SERVOMIN);
+    Serial.println("Searching for index");
+    while (digitalRead(indexPin) == LOW) {
+        Serial.print(".");
+        delay(50);
+    }
+    
+    // Stop motor, and reset encoder position
+    pwm.setPWM(pwmChannel, 0, 0);
+    encoder.write(0);
+
+    delay(1000);
+}
