--- src/motors.h
@@ -1,13 +1,0 @@
-#ifndef MOTORS_H
-#define MOTORS_H
-
-#include <BasicLinearAlgebra.h>
-#include <Adafruit_PWMServoDriver.h>
-#include <Encoder.h>
-
-void moveMotor(Adafruit_PWMServoDriver &pwm, uint8_t motor_no, float degrees);
-void moveMotors(Adafruit_PWMServoDriver &pwm, BLA::Matrix<3> delta_angles);
-
-BLA::Matrix<3> getMotorAngles();
-
-#endif
